HYD-5419 Maybe: implement Maybe datatype
Implement the Maybe datatype.

Maybe is a datatype that represents a value that may or may not be
there.

It is useful to explicitly model where a possible null type may occur.

Once we wrap a value in Maybe, all future computations will be skipped
if they are null (or undefined).

  - Create initial metadata files.
  - Create Maybe.
  - Create tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>