HYD-6751 Update Maybe
The Maybe type used to be a class that was exported directly.

In addition, it wasn't parameterized so the containing type was opaque.

Update Maybe to remove the class as a direct export; we will use module
level functions instead.

In addition, parmeterize the class so we can operate over the contained
type.

  - Update deps.
  - Update dotfiles.
  - Change Maybe to use static functions.

Signed-off-by: Joe Grund <joe.grund@intel.com>