Update (#2)
* HYD-5709 add coverage to maybe

Now that we have code coverage generation in CI,
we can build upon this metric. We need our transpiled,
ES6 code to also be covered accurately.

We are going to need one for the GUI,
and one for NodeJS projects.

 - Updated for coverage.

Signed-off-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>

* CR updates

Signed-off-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>

* Tests passing locally, coverage too

Signed-off-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>

* HYD-6751 Update Maybe

The Maybe type used to be a class that was exported directly.

In addition, it wasn't parameterized so the containing type was opaque.

Update Maybe to remove the class as a direct export; we will use module
level functions instead.

In addition, parmeterize the class so we can operate over the contained
type.

  - Update deps.
  - Update dotfiles.
  - Change Maybe to use static functions.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Version 2.0.0

4629928 HYD-6751 Update Maybe
952669e Tests passing locally, coverage too
03e0149 CR updates
367dfdb HYD-5709 add coverage to maybe

* HYD-6780 Maybe: Add matchWith method

Add a matchWith method to Maybe.

It will serve as a case expression matcher for Just and Nothing.

example:

```
maybe.matchWith(
  {
    Just:(x) => `ba${x}`,
    Nothing:() => 'bap'
  },
  maybe.of('p')
)
```

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Version 2.1.0

5241c26 HYD-6780 Maybe: Add matchWith method

* HYD-7044 maybe: stop compiling to ES6

Stop compiling maybe to ES6. We can use downstream repos to do this.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* 2.1.1

* update to work with flow changes

Signed-off-by: Joe Grund <joe.grund@intel.com>

* 2.1.2

* Remove fp.

Remove the fp lib from maybe.

We should start looking at moving away from curry
due to difficulties of flow parsing types from it correctly.
Instead, we can use Function.prototype.bind.

  - Remove fp dep.
  - Remove .flow file and use source for typing.
  - Bump deps.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* 3.0.0

* Update to latest.